% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_glmnet_single.R
\name{anova_glmnet_single}
\alias{anova_glmnet_single}
\title{Assign ANOVA type-II -log10 p-values to one submodel of a sparse \code{glmnet} model}
\usage{
anova_glmnet_single(
  X,
  y,
  pcs = NULL,
  obj_cv = NULL,
  beta = NULL,
  index = "min",
  ret_sparse = FALSE
)
}
\arguments{
\item{X}{The genotype matrix.
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.}

\item{y}{The trait vector.
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.}

\item{pcs}{The PC (eigenvector) matrix (optional).
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.
Unlike genotypes, PCs are not given p-values.}

\item{obj_cv}{Optional, the cross-validation object produced by \code{\link[=glmnet_pca]{glmnet_pca()}} with \code{cv = TRUE} (must be class "cv.glmfit").
Either this or \code{beta} must be provided.}

\item{beta}{Optional, the sparse matrix of coefficients (component \verb{$beta}) of the \code{glmnet} object (output of \code{\link[=glmnet_pca]{glmnet_pca()}} with \code{cv = FALSE}, or component \verb{$glmnet.fit$beta} if \code{cv = TRUE}).
Either this or \code{obj_cv} must be provided.}

\item{index}{The index for the desired \code{lambda} penalty factor, which implicitly chooses the submodel to analyze.
This can be a numeric index corresponding to a column of \code{beta}.
Alternatively, if \code{obj_cv} is provided, \code{index} can be "min" (default, selects \code{obj_cv$index[1]} as the actual index, which corresponds to the \code{lambda} that minimized the mean cross-validation error) or "1se" (\code{obj_cv$index[2]}, which is the largest \code{lambda} such that error is within 1 standard error of the minimum).
See \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} for more information on "min" and "1se" definitions.}

\item{ret_sparse}{Logical that controls return value (see that).}
}
\value{
If \code{ret_sparse = FALSE} (default), returns a complete vector of scores (-log10 p-values) for every locus in \code{X}, with zeroes for all loci with zero coefficients.
For loci with non-zero coefficients, p-values are calculated using \code{\link[=anova2]{anova2()}}, see that for more details.
If \code{ret_sparse = TRUE}, returns a list of indexes and scores corresponding only to the loci with non-zero coefficients.
}
\description{
Given a sparse \code{glmnet} model (not ridge regression), here we assing ANOVA type-II -log10 p-values to a single submodel obtained by specifying providing the cross-validation \code{glmnet} object and specifying index as "min" or "1se" (see \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} return object), or alternatively only the \code{beta} matrix (sparse coefficients) and the numeric index of the \code{lambda} penalty factor.
To achieve this, each set of selected loci is fit again to the original data without penalization.
}
\examples{
\dontrun{
# version with cross-validation object `obj_cv` (recommended)
# defaults to selecting model with lowest cross-validation error (`index = "min"`)
scores <- anova_glmnet_single( X, y, pcs, obj_cv = obj_cv )

# version with beta matrix and desired index
scores <- anova_glmnet_single( X, y, pcs, beta = beta, index = 50 )
}

}
\seealso{
\code{\link[=glmnet_pca]{glmnet_pca()}}, particularly option \code{cv = TRUE}, for obtaining cross-validation objects with PCs.

\code{\link[=anova_glmnet]{anova_glmnet()}} for a version that calculates scores for all models (all lambdas), though it is much slower and not generally recommended.

\code{\link[=anova2]{anova2()}} for additional details and data restrictions.

\code{\link[=scores_glmnet]{scores_glmnet()}} for a different way of scoring/raking variants.
}
