% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_single.R
\name{anova_single}
\alias{anova_single}
\title{Assign ANOVA type-II -log10 p-values to a sparse genetic model}
\usage{
anova_single(X, y, indexes, pcs = NULL, ret_sparse = FALSE)
}
\arguments{
\item{X}{The genotype matrix.}

\item{y}{The trait vector.}

\item{indexes}{Indexes of the loci to fit.}

\item{pcs}{The PC (eigenvector) matrix (optional).
Unlike genotypes, PCs are not given p-values.}

\item{ret_sparse}{Logical that controls return value (see that).}
}
\value{
If \code{ret_sparse = FALSE} (default), returns a complete vector of scores (-log10 p-values) for every locus in \code{X}, with zeroes for all loci with zero coefficients.
For loci with non-zero coefficients, p-values are calculated using \code{\link[=anova2]{anova2()}}, see that for more details.
If \code{ret_sparse = TRUE}, returns a list of indexes and scores corresponding only to the loci with non-zero coefficients.
}
\description{
Given a sparse model defined by explicit locus indexes, here we assing ANOVA type-II -log10 p-values to these loci.
To achieve this, loci coefficients are fit to the data as multiple linear regression, without penalization.
}
\examples{
\dontrun{
scores <- anova_single( X, y, indexes, pcs )
}

}
\seealso{
\code{\link[=glmnet_pca]{glmnet_pca()}}, particularly option \code{cv = TRUE}, for obtaining cross-validation objects with PCs.

\code{\link[=anova_glmnet_single]{anova_glmnet_single()}} for scores on a single \code{glmnet} model, instead of specifying indexes explicitly as here.

\code{\link[=anova_glmnet]{anova_glmnet()}} for scores for all \code{glmnet} models (all lambdas), though it is much slower and not generally recommended.

\code{\link[=anova2]{anova2()}} for additional details and data restrictions.

\code{\link[=scores_glmnet]{scores_glmnet()}} for a different way of scoring/raking variants.
}
