% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_glmnet.R
\name{scores_glmnet}
\alias{scores_glmnet}
\title{Score loci by appearance in parse \code{glmnet} model}
\usage{
scores_glmnet(beta)
}
\arguments{
\item{beta}{The matrix of coefficients (component \verb{$beta}) of the \code{glmnet} object.}
}
\value{
The score vector.
Each locus receives a score between 0 and \code{k} (number of columns of \code{beta}).
Loci that were never selected have zero scores.
Loci first selected in column \code{i} have scores of \code{k + 1 - i}, so loci selected earlier have higher scores.
}
\description{
The \code{glmnet} package does not provide a way to rank variants.
Here is implemented one possible ranking based on when a variable is first selected as the penalty factor \code{lambda} of \code{glmnet} is varied.
\code{glmnet} precomputes coefficients at a series of \code{lambda} knots, which are used for this score calculation.
This approach requires sparsity (does not work for ridge regression).
}
\examples{
\dontrun{
scores <- scores_glmnet( beta )
}

}
\seealso{
\code{\link[=anova_glmnet]{anova_glmnet()}} for a different way of scoring/raking variants.
}
