% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_glmnet.R
\name{anova_glmnet}
\alias{anova_glmnet}
\title{Assign ANOVA type-II p-values to every submodel of a sparse \code{glmnet} model}
\usage{
anova_glmnet(beta, X, y, pcs = NULL)
}
\arguments{
\item{beta}{The matrix of coefficients (component \verb{$beta}) of the \code{glmnet} object.}

\item{X}{The genotype matrix.
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.}

\item{y}{The trait vector.
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.}

\item{pcs}{The PC (eigenvector) matrix (optional).
Same as was used in \code{\link[=glmnet_pca]{glmnet_pca()}}.
Unlike genotypes, PCs are not assigned p-values.}
}
\value{
A type-II ANOVA p-value matrix with the same dimensions as \code{beta}.
Zero coefficients (unselected variables) are assigned p-values of 1.
For each column, selected variables are assigned p-values using \code{\link[=anova2]{anova2()}}.
}
\description{
Given a sparse \code{glmnet} model (not ridge regression), here we assing ANOVA type-II p-values to every submodel obtained by varying the \code{lambda} penalty factor (i.e. each column of the \verb{$beta} component matrix of the \code{glmnet} object).
To achieve this, each set of selected loci is fit again to the original data without penalization.
}
\examples{
\dontrun{
pvals <- anova_glmnet( beta, X, y, pcs )
}

}
\seealso{
\code{\link[=anova2]{anova2()}} for additional details and data restrictions.

\code{\link[=scores_glmnet]{scores_glmnet()}} for a different way of scoring/raking variants.
}
