% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_pca.R
\name{glmnet_pca}
\alias{glmnet_pca}
\title{Fit a polygenic model with \code{glmnet} and principal components}
\usage{
glmnet_pca(X, y, pcs = NULL, ...)
}
\arguments{
\item{X}{The genotype matrix.
It must be oriented with loci along rows and individuals along columns, which agrees with other genetics packages by the main author, although this is transposed from what \code{glmnet} and other regression models expect (even \code{lm}).}

\item{y}{The trait vector.
It must have length equal to the number of individuals.}

\item{pcs}{The PC (eigenvector) matrix (optional).
It must have individuals along the rows and dimensions along the columns.
Unlike \code{X}, predictors in this \code{pcs} matrix, are not penalized, and their coefficients are not included in the output.}

\item{...}{Additional parameters passed to \code{glmnet}.
Cannot include \code{penalty.factor}, which is set internally so that loci are equally penalized and \code{pcs} are unpenalized.}
}
\value{
A \code{glmnet} object.
Here it includes coefficients for every locus in \code{X} (can be sparse) but excludes coefficients for the \code{pcs} elements.
}
\description{
This function makes it easy to fit a LASSO or Elastic Net model while including principal components or other fixed effects that are not part of the variable selection process and are not penalized.
This resembles in spirit and functionality the \code{snpnet} function of the same package, except that one requires the genetic data to be written into a plink2 file whereas this one requires inputs that are ordinary R matrices.
}
\examples{
\dontrun{
obj <- glmnet_pca(X, y, pcs)
}

}
